move CSV and Spec Reader support up to QL
