+++ x-pack/plugin/sql/qa/server/multi-node/src/javaRestTest/java/org/elasticsearch/xpack/sql/qa/multi_node/GeoJdbcCsvSpecIT.java
@@ -8,7 +8,8 @@
 package org.elasticsearch.xpack.sql.qa.multi_node;
 
 import org.elasticsearch.xpack.sql.qa.geo.GeoCsvSpecTestCase;
-import org.elasticsearch.xpack.sql.qa.jdbc.CsvTestUtils.CsvTestCase;
+
+import static org.elasticsearch.xpack.ql.CsvSpecReader.CsvTestCase;
 
 public class GeoJdbcCsvSpecIT extends GeoCsvSpecTestCase {
     public GeoJdbcCsvSpecIT(String fileName, String groupName, String testName, Integer lineNumber, CsvTestCase testCase) {
