+++ x-pack/plugin/sql/qa/server/security/build.gradle
@@ -7,6 +7,7 @@ dependencies {
   testImplementation(project(':x-pack:plugin:sql:qa:server')) {
     transitive = false
   }
+  api project(xpackModule('ql:test-fixtures'))
 }
 
 Project mainProject = project
