+++ x-pack/plugin/sql/qa/server/single-node/src/javaRestTest/java/org/elasticsearch/xpack/sql/qa/single_node/JdbcCsvSpecIT.java
@@ -9,12 +9,12 @@ package org.elasticsearch.xpack.sql.qa.single_node;
 import com.carrotsearch.randomizedtesting.annotations.ParametersFactory;
 
 import org.elasticsearch.xpack.sql.qa.jdbc.CsvSpecTestCase;
-import org.elasticsearch.xpack.sql.qa.jdbc.CsvTestUtils.CsvTestCase;
 
 import java.util.ArrayList;
 import java.util.List;
 
-import static org.elasticsearch.xpack.sql.qa.jdbc.CsvTestUtils.specParser;
+import static org.elasticsearch.xpack.ql.CsvSpecReader.CsvTestCase;
+import static org.elasticsearch.xpack.ql.CsvSpecReader.specParser;
 
 public class JdbcCsvSpecIT extends CsvSpecTestCase {
