+++ x-pack/plugin/sql/qa/server/src/main/java/org/elasticsearch/xpack/sql/qa/jdbc/SpecBaseIntegrationTestCase.java
@@ -10,29 +10,23 @@ import org.apache.logging.log4j.Logger;
 import org.elasticsearch.client.Request;
 import org.elasticsearch.client.ResponseException;
 import org.elasticsearch.client.RestClient;
-import org.elasticsearch.common.Strings;
-import org.elasticsearch.xpack.ql.TestUtils;
+import org.elasticsearch.xpack.ql.SpecReader;
 import org.junit.AfterClass;
 import org.junit.Before;
 
-import java.io.BufferedReader;
 import java.io.IOException;
 import java.net.URL;
 import java.sql.Connection;
 import java.sql.ResultSet;
 import java.sql.SQLException;
 import java.sql.Statement;
-import java.util.ArrayList;
-import java.util.LinkedHashMap;
 import java.util.List;
 import java.util.Locale;
-import java.util.Map;
 import java.util.Objects;
 import java.util.Properties;
 import java.util.TimeZone;
 
-import static java.util.Collections.emptyList;
-import static org.elasticsearch.xpack.ql.TestUtils.pathAndName;
+import static org.elasticsearch.xpack.ql.SpecReader.Parser;
 import static org.elasticsearch.xpack.sql.qa.jdbc.JdbcTestUtils.JDBC_TIMEZONE;
 
 /**
@@ -158,75 +152,6 @@ public abstract class SpecBaseIntegrationTestCase extends JdbcIntegrationTestCas
         URL source = SpecBaseIntegrationTestCase.class.getResource(url);
         Objects.requireNonNull(source, "Cannot find resource " + url);
 
-        return readURLSpec(source, parser);
-    }
-
-    protected static List<Object[]> readScriptSpec(List<URL> urls, Parser parser) throws Exception {
-        List<Object[]> results = emptyList();
-        for (URL url : urls) {
-            List<Object[]> specs = readURLSpec(url, parser);
-            if (results.isEmpty()) {
-                results = specs;
-            } else {
-                results.addAll(specs);
-            }
-        }
-
-        return results;
-    }
-
-    private static List<Object[]> readURLSpec(URL source, Parser parser) throws Exception {
-        String fileName = pathAndName(source.getFile()).v2();
-        String groupName = fileName.substring(0, fileName.lastIndexOf("."));
-
-        Map<String, Integer> testNames = new LinkedHashMap<>();
-        List<Object[]> testCases = new ArrayList<>();
-
-        String testName = null;
-        try (BufferedReader reader = TestUtils.reader(source)) {
-            String line;
-            int lineNumber = 1;
-            while ((line = reader.readLine()) != null) {
-                line = line.trim();
-                // ignore comments
-                if (line.isEmpty() == false && line.startsWith("//") == false) {
-                    // parse test name
-                    if (testName == null) {
-                        if (testNames.keySet().contains(line)) {
-                            throw new IllegalStateException(
-                                "Duplicate test name '"
-                                    + line
-                                    + "' at line "
-                                    + lineNumber
-                                    + " (previously seen at line "
-                                    + testNames.get(line)
-                                    + ")"
-                            );
-                        } else {
-                            testName = Strings.capitalize(line);
-                            testNames.put(testName, Integer.valueOf(lineNumber));
-                        }
-                    } else {
-                        Object result = parser.parse(line);
-                        // only if the parser is ready, add the object - otherwise keep on serving it lines
-                        if (result != null) {
-                            testCases.add(new Object[] { fileName, groupName, testName, Integer.valueOf(lineNumber), result });
-                            testName = null;
-                        }
-                    }
-                }
-                lineNumber++;
-            }
-            if (testName != null) {
-                throw new IllegalStateException("Read a test without a body at the end of [" + fileName + "].");
-            }
-        }
-        assertNull("Cannot find spec for test " + testName, testName);
-
-        return testCases;
-    }
-
-    public interface Parser {
-        Object parse(String line);
+        return SpecReader.readURLSpec(source, parser);
     }
 }
